Do not check global variables' initialization
There are several problems with checking global variables.
First of all, we would need a CFG for the initialization expression to
properly support the check. Right now, the clang AnalysisDeclContext
does not support this. While it is possible to add, I think it does
not give us much value.

The only thing we can check according to the analysis rulses whether
we init the global with (static) or (null). But we have no local
variables in the scope, and if we return a local from a function we
will diagnose that in the callee. The only value we could bring is
the null check for not nullable types. I think this could be
addressed later if we find that it is worth the cost of the
implementation.