Remove the last pass over the CFG to generate reports.
Now we report everything as we do the analysis. This has a lot
of advantages:
* Reporter is no longer a pointer, less null checks, less complexity
  and less possible sources of errors.
* Better performance, one less pass over the cfg.
* If loops re involved, we will emit the debug psets in each iteration.
  This helps us understand what is going on and makes it possible to
  write more rigorous tests.

It should be save to report an error as soon as we see it. We should
only add new elements to the pset in each iteration, never remove.
So it is safe to warn in the first iteration we see null or invalid
in a pset where it should not be. In case we start to se duplicate
warnings, we could still deduplicate warnings within the reporter
rather than making the PsetBuilder more complex.