Admit that iterators are nullable
Also, since we did not warn when a nullable type was null during a call
I added the requirement that *this should never point to null. This way
we get a warning for *iter when iter points to null, even if iter itself
is nullable. This might be too eager though. It is perfetly reasonable
to do some method calls on a null Pointer, like equality. In case of
unique_ptr (even though it is an owner), it is reasonable to call some
methods when it is null, but it is still error to call other methods.

What is the good solution here?