Make IgnoreLValueToRValue parameter honest.
Setting IngorLValueToRValue casts to true also ignored null to pointer
casts. I do not have a testcase where it caused problems but it did not
break anything either. I did the change anyways, so the methods reflects
the original intention. Feel free to revert, if the intention was
something else.