Fix crashes with lambda captures
Interestingly, sometimes we do not really see references to the fields
of the closure object, rather we see DeclRefExpr referring to the
captured variable directly. Let's assume that the lifetime of those are
static. We should diagnose dangling lambdas during the caller analysis,
so this should be a safe bet for the callee analysis. Also, it looks
like we cannot really query the list of captures yet, the operator() is
analyzed before the capture list is populated by Sema.