Remove a TODO comment
I think it is perfectly reasonable to build our own CFG for lifetime
analysis since the CFG we use likely to be different from the one that
other analyses are using (since we are doing deeper analysis).

Looking at the performance numbers we have so far building our own
should not be that bad.