Add support for deref locations in Variable
This change is a prerequisite for a lot of features like proper
post-condition violations and aggregate support. Implementing this also
showed us references are special and they need a lot of
workaround/special casing. It would be great to somehow abstract away
from them.

Calculating the type of a variable is no longer that cheap, maybe it
would be worth to store an always up to date version?

That way we could also make this and temporary variables typed in
the representation.

This fixes #22.