Allow expansion statements to find uninstantiated decls
This cleans up some logic surrounding instantiations, for instance, we no longer
need to explicitly map the range variable to itself. This additionally results
in the resolution of the issue where in dependent contexts we would get two
copies of the range variable.

Additionally this commits cleans up some unnecessary logic, and removes "tuple"
wording from the implementation where it is not relevant.