Implemented reflexpr(::) as well as preservation of namespace qualifiers
The majority of this commit is a rework of namespace handling to use
a pointer union of TranslationUnitDecl and NamespaceDecl to represent
namespaces before constexpr evaluation.

This is necessary as clang lacks a namespace object to represent the
global namespace. Thus, this uses the TranslationUnitDecl as said
object.

This has been encapsulated in an alias type ReflectedNamespace.
This should be trivially replacable should clang's global namespace
implementation change to use a namespace object.

The NamespaceName class now has all objects dynamically allocated
to allow for space for NestedNameSpecifiers. One noted caveat in
this implementation is that now all namespace declaration reflections
will result in a -- small -- additional allocation per reflexpr.
This tradeoff is made for simplification of the implementation,
at what is expected to be a relatively low cost. The change has
no impact on the performance of namespaces reflected internally
from APValues.