Implemented reflexpr(::) as well as preservation of namespace qualifiers
The majority of this commit is a rework of namespace handling to use
a pointer union of TranslationUnitDecl and NamespaceDecl to represent
namespaces before constexpr evaluation.

This is necessary as clang lacks a namespace object to represent the
global namespace. Thus, this uses the TranslationUnitDecl as said
object.

This has been encapsulated in an alias type ReflectedNamespace.
This should be trivially replacable should clang's global namespace
implementation change to use a namespace object.