Add a missing test and fix a bug.
Calling a method on a "null" state pointer considered an error.
(Is this inteded? Why do we check null-nes for this? Shouldn't most
pointer methods handle this case?)

So for now I added an expception, conversions to bools or raw pointers
can be invoked on null Pointers as they should definitely handle that
case by returning false or nullptr.