Some experiments parsing the gsl::lifetime attribute
There are several challenges following the paper:

The paper has two possible positions for gsl::lifetime. Are we supposed to
support both or one of them?
 *  [[gsl::lifetime(p,{o})]] f(int* p, string& o)
 * void f(int* p, int* p2 [[gsl::lifetime({p})]])

Also, what to do about late parsing?
void basic_lateparse(int *a [[gsl::lifetime(b)]], int *b);

The problem is that the identifier of b is introduced after the argument.
While both Clang and the standard has the notion of late parsing it is only
required inside the classes to support mentioning identifiers of member. Thus,
supporting the syntax above either needs additional non-standard facilities
from the compiler or we need to introduce our own parsing. Since we only need
to parse identifiers and some symbol for deref locations doing our own parsing
should not be that hard but I wonder whether it is a good idea to require that
from the implementations.

In the paper I found the following spelling to annotate return values:
int* f(int* p, int* p2, string o) [[gsl::lifetime{return,{p,o})]]
The return is not really an identifier but a keyword so this will definitely
require the implementations to have custom parsing rules. An alternative would
be to somehow piggybacking on the ensure annotation.

I also saw the following example:
[[gsl::lifetime(*a,{*b}]] [[gsl::lifetime(*b,{*a}]] swap(T& a, T& b)
 
I think Clang does not really support having multiple annotations of the same
type attached to a node so I do not think this being feasible.

Possible workaround:
[[gsl::lifetime(*a, {*b}, *b, {*a})]] swap(T& a, T& b)

Note: this is not related to parsing, just a possible source of confusion. The
annotation above is actually a postcondition. Annotations earlier were
preconditions. So gsl::lifetime can both be a post or a precondition depending
on the argument it is applied to being categorized as input or output. One can
get used to it but it might be confusing at first. Should we consider two
separate annotations one for preconditions and one for postconditions?
This could subsume the gsl::out annotation?