More experiments with the gsl::lifetime attribute
Note that this is merely an experiment and gsl::lifetime annotated code
does not expected to work well yet. Right now we have some basic custom
parsing for the attribute and the argument of the attribute is a string
literal. The reason is that the attribute arguments are represented as
a union of Identifiers and Exprs and I did not find a good way to get an
argument as a Token stream. A token stream would be the best
representation for the kinds of annotations required by the paper.

Also, I use indexes to refer to the parameters. This is a great approach
because this way the same attribute is compatible with all the
redeclarations, even if the redeclarations have different parameter
names.

One of the big problems of gsl::lifetime attribute it is really hard to
know whether it should result in a precondition or a postcondition
check. The attribute on an input argument results in a precondition
check and on an output attribute it results in a postcondition check. So
we need to classify inputs/outputs first in order to know what to do
with a gsl::lifetime annotation. Moreover, we might have entities that
are both inputs and outputs like *this in an assignment. How are we
supposed to treat gsl::lifetime annotation in those cases?