Explore the possibility of parsing pre/post exprs
One major problem of the current design was that it is not apparent
whether a gsl::lifetime annotation is a pre- or a postcondition. There
is a proposal, however, about pre- and postconditions: contracts!

I started to explore an interesting avenue: building a DSL that can be
used as a contract. The essential goal would be to be able to write
exprs written in this DSL in the contract attributes and make the
lifetime analysis understand that. This way the user of the lifetime
analysis could get some runtime verification for free.

This commit introduces the first experimental bits, this is far from
being polished.