Make the precondition assumption entirely annotation driven
Right now all the input pset information is derived from the
annotations. Unfortunately, I had to do a lot of refactoring to make the
solution viable. This is not 100% working, but it is getting close. One
of the biggest change is that we no longer have order and deref
separately. The rationale is that we do not want the user to learn both
concepts (when they write the annotations and a pointes is pointing into
an Owner it would be great to have the same syntax as a pointer pointing
'into' another pointer, so *unique_ptr and *raw_ptr can be written in
both cases). The internal representation should match this mental model.

The only place where this does not really work out yet is invalidating
owners through pointers. This should be easy to fix.

We also lost some notes temporarily, but I also plan to add them back
soon.