Remove some questionable optimization
Checking if two psets are equal has similar complexity as actually
merging them and merging to equal psets should not change their value.
Thus, always merge PSets without checking for equality.

We used to have an optimization if o' and o'' is both in a PSet it is ok
to only store one of them. I removed this optimization earlier during a
refactoring. Now I remove the TODO as well. The reason is that o' and
o'' are very likely to have separate types and most PSets should contain
pointers that have the same type (unless annotation tells otherwise).
So it might not be worth to optimize for this rare case and pay the cost
in the more common case.