Add sketch to derive default postconditions
At first sight there is a large amount of code duplication. I expect
that duplication to dissappear once the analysis starts to use the
derived attributes. This patch does not introduce the actual derivation
yet as I just realized I cannot use the
LifetimeContractAttr::PointsToLoc representation as it is tied to the
declaration. I should use an index there as well. Once it is resolved I
plan to continue this work in follow-up commits.