Support return values in contracts
Right now return values are expressed as "temporaries". This works for
now as temporaries do not have names so cannot appear in contracts.

In the long run we need to think about a contracts representation
carefully as it is very different from the PSet representation in the
analysis. The key differences are:

* Contract representation must be independent of the function declarations.
* Contract representation must have an entry for the return value which
  has no corresponding declaration at all.
* Contract representation never refers to temporaries.
* Contract representation always have a parameter or the return value as
  its base while Variables during the analysis might have other local
  variables
* PMap representation never have an entry for the return value. We
  calculate psets only for params, locals, exprs.