Move the representation of contracts and psets closer
This helps a bit with code reuse. Note that, since the ExprCXX and
DeclCXX headers are including Attr.h, it is not possible to refer to any
C++ specific node in Attr.h in a way that require a complete type. The
pointer union requires complete types.

This commit also makes it possible to query the type of a Variable
representing this variable. This slightly changes the representation.