Nulls can no longer propagate from contracts.
For cases like:
int *p = f(&x);

We want p not to contain null as null was not there among the inputs.

In the future we also want users to be able to express that null should
be among the outputs regardless of the inputs. This is not supported for
now.