Treat forwarding references as if they were annotated lifetime_const
Unfortunately, I did not find a trivial way to tell if a template
parameter is a forwarding reference. So I ended up adding the annotation
in the Sema module.

While working on this feature I also noticed and fixed the default
contract generation did not consider the lifetime_const annotations.

TODO: the lifetime_const annotation is still using the old syntax (no
expression argument).