Attempt to fix some false positives
Unfortunately, I do see a tricky scenario.

Basically, we are interested to see what is the initializer of a
gsl::Pointer. If we initialize a gsl::Pointer with a local or a
temporary gsl::Pointer that is fine, it is analogous to copying the
pset. If we initialize a gsl::Pointer with a local or temporary
gsl::Owner, that is a bug for sure, we have to warn.

The tricky question is, what to do if we initialize a gsl::Pointer using
an unannotated type? Should we assume that the pointer points to that
type? If we do not assume that, we will end up having false negatives.
If we do assume that and the user does not annotate some types, we will
get false positives.