Fix a false positive, but...
So the problem why we did not warn for the following code:
std::optional<std::vector<int>> getOptVec();

for (int value : *getOptVec())
                 ^~~~~~~~~~~~  <- Here!

Because we was not able to derive the output for optional::operator*.
Since the const T &operator*() const overload came before the non-const
one we infered the DerefType to be const T for optional<T>. And since
we call the non-const overload in the code above due to the conversion
rules none of the inputs matched the outputs.

This shows us how fragile the type inference can be in this regard,
depending on which overload do we find first we might end up with either
T or const T for the DerefType.

To circumvent the issue for std::optional, I prioritized the value_type
typedef over the return type of special methods. This will solve this
particular issue but it is not a universal solution for user defined
types.