Add post dominator based warning filtering.
Note that, this is more of a workaround than a proper solution.
There are some test cases to demonstrate why.

The correct solution would be to both overestimate and underestimate the
points to sets. Currently we only overestimate them, so we have an upper
bound on the points to sets. What we want to do instead, to also have a
lower bound, and warn in two specific cases:
* If null/invalid is in the lower bound of the points-to set.
* If null/invalid is in the upper bound of the points-to set and the
dereference post-dominates the source.

I beleive it would be possible to maintain both upper and lower bounds
relatively cheaply while running the same analysis and it would give
much better results compared to the current "workaround" for showing
only filtered results.