Fix a divergence problem with owners.
We have a monotonic analysis in a sense that each PSet can only grow
during the abstract interpretation. A such a growing analysis is only
guaranteed to terminate when we have an upper bound on the size of the
program state. The worst case scenario is when each pointer can point to
every possible memory location, so we are bounded if and only if there
is a bound on the number of abstract memory locations.

We did not have such a bound since we always conjured a new memory
location each time we dereferenced an owner. This way it was possible to
trigger divergence. It was extremely rare (only one function triggered
it in whole Fuchsia), but still it is a bug. To get a bound on the
possible memory locations I added a hard limit on the number of memory
locations we conjure from an owner.