Reimplement the lifetime analysis filters.
We want to prove the following:
* there is dead code in the function, or
* there is a feasible path from null/invalid to deref.

If we assume there is no dead code, there must be a feasible path between
the two nodes when they are in a domination relation.

There is no way to tell which paths are feasible and which paths are
not. So in order to not to bother users with false positive warnings, we
will only warn if none of the paths between the null/invalid and the
deref are safe, i.e. none of them overwrites the null/dangling ptr.