Introduce a proper WorkList algorithm
On the tests we have we visit process the CFGBlocks the same number of
times. We do, however, less work overall, since we do not enqueue the
successors of a node when no processing required for it. Previously we
always iterated over the whole CFG, now we will only do the fixed point
iteration on the nodes that actually has a changed state.

Moreover, this approach is more idiomatic to dataflow analysis and this
is the pattern that the rest of the flow sensitive checks are following.
As a nice side effect, BlockContext now only contains data that is
related to the actual analysis, and all the traversal related date is
separated from that.

This will also help with the upstreaming efforts as clang developers
want us to contribute reusable pieces. This will help in the long run to
factor the traversal out to be reusable.