Better support for std::list, std::forward_list
In fact, most of the methods of these classes will not do any
invalidation. This does not really solve all the problems, as some
methods, like splice, can do ownership transfers which are currently not
supported.

Also note that, for example, erase will invalidate only some iterators,
those that point to the erased element. The lifetime analysis has no
knowledge about which elements the iterators are pointing at. Should we
just assume that all remain valid? (This is the behavior that the
current patch implements.)