Reduce delta with master
We added the include in DeclTemplate.h. I moved it to
some cpp files hopefully reducing compile times when we
tinker with attributes.
The other include in SemaDecl.cpp was just redundant.

We probably no longer need the workaround in the lit config?
I can run the tests without it.

Other small changes to reduce delta.