Fixed suppression of constexpr evaluation errors for metaprograms
This fixes an issue where all constexpr uninitialized value
diagnostics were accidentally being suppressed to allow for
metaprogram decls to effectively return void constexpr evaluation
values.

Additionally, this includes a fix for source location information
for the __constexpr_decl generated function, although,
we should likely look for ways to suppress this in the diagnostics
as it itself, may be a source of confusion.