Make Fragments always meta::info type
Fragments are now represented via a dedicated APValue fragment type.
This type is a variant of a reflection type, seperated out for ease of
implementation, and efficiency reasons.

Fragment expressions have been updated to better conform to modern
clang architecture, with a new Create factory method, trailing objects
to store the capture expressions, and the use of a bits class in
Stmt.h for more compact memory representation.

Additionally, fragment expressions no longer use an explicitly created
initializer, or a closure class. This greatly reduces complexity, as
this closure type no longer has to be synthesized, and we're able to
directly evaluate and store captured values into a fragment APValue.