Rewrite how parameter types are handled.
This makes input parameters work, more or less, and generate valid code.
Overload resolution and name mangling aren't implemented yet.

The approach is to do almost everything in terms of a parameter's
adjusted type: conversions, initialization, code generation. Whenever we
see a parameter type or an implicit cast to parameter type, we use the
adjusted type or ignore the cast, respectively.