Merge fix: Fixed compilation of table gened files for PropertiesBase.td
Specifically, this fix allows PackSplices to appear in the generated
code, and for it to compile. However, said code will ultimately fail
as PackSplices should not be copied.

Further work needs to be done to determine the best resolution for this issue
which became visible after upstream gave llvm::TrailingObjects a deleted
copy constructor exposing the issue at compile time.