Merge fix: Restore getSubExprAsWritten to original state
Previously we modified getSubExprAsWritten to prevent
the ComplexRemove TreeTransform class from breaking uses
of consteval temporaries (via removing the wrapping ConstantExpr).

However, there are API level tests that want to gaurantee that
the wrapping ConstantExpr is removed by getSubExprAsWritten.

Instead, fix this by allowing ComplexRemove to override
a new TransformSubExprAsWritten, replacing the problematic call to
getSubExprAsWritten, with one to getSubExpr. This is likely
still a bit of a hack, and finding a better solution
should still be considered.