Start rewriting splice parsing and semantics.
Most splices are going to be parsed as part of a nested-name-specifier.
Make sure we can parse those fully in that context, creating annotated
tokens for the splice.

We can evaluate the spliced expression at the point where we expect the
splice to become its inserted entity.

So far this (seems) to work for splicing types (typename [:refl:]).